refactor: rename CEF input handler
