+++ src/cef/ui/uicefwebview.cpp
@@ -27,8 +27,8 @@
 #include <framework/core/eventdispatcher.h>
 #include <framework/graphics/graphics.h>
 #include <framework/graphics/image.h>
-#include <framework/platform/platformwindow.h>
 #include "cef_client.h"
+#include "cef_inputhandler.h"
 #ifdef _WIN32
 #include <io.h>
 #else
@@ -36,7 +36,6 @@
 #endif
 #include <string>
 #include <vector>
-#include <unordered_map>
 #include <algorithm>
 #include <cstring>
 #include <memory>
@@ -156,142 +155,6 @@ void* UICEFWebView::s_eglDisplay = nullptr;
 void* UICEFWebView::s_eglContext = nullptr;
 #endif
 
-static int getCefModifiers()
-{
-    int mods = 0;
-
-    int keyMods = g_window.getKeyboardModifiers();
-    if (keyMods & Fw::KeyboardCtrlModifier)
-        mods |= EVENTFLAG_CONTROL_DOWN;
-    if (keyMods & Fw::KeyboardShiftModifier)
-        mods |= EVENTFLAG_SHIFT_DOWN;
-    if (keyMods & Fw::KeyboardAltModifier)
-        mods |= EVENTFLAG_ALT_DOWN;
-
-    if (g_window.isMouseButtonPressed(Fw::MouseLeftButton))
-        mods |= EVENTFLAG_LEFT_MOUSE_BUTTON;
-    if (g_window.isMouseButtonPressed(Fw::MouseRightButton))
-        mods |= EVENTFLAG_RIGHT_MOUSE_BUTTON;
-    if (g_window.isMouseButtonPressed(Fw::MouseMidButton))
-        mods |= EVENTFLAG_MIDDLE_MOUSE_BUTTON;
-
-    return mods;
-}
-
-static int translateKeyCode(uchar key)
-{
-    static const std::unordered_map<uchar, int> vkMap = {
-        {Fw::KeyBackspace, 0x08}, // VK_BACK
-        {Fw::KeyTab, 0x09},       // VK_TAB
-        {Fw::KeyEnter, 0x0D},     // VK_RETURN
-        {Fw::KeyEscape, 0x1B},    // VK_ESCAPE
-        {Fw::KeyPageUp, 0x21},    // VK_PRIOR
-        {Fw::KeyPageDown, 0x22},  // VK_NEXT
-        {Fw::KeyEnd, 0x23},       // VK_END
-        {Fw::KeyHome, 0x24},      // VK_HOME
-        {Fw::KeyLeft, 0x25},      // VK_LEFT
-        {Fw::KeyUp, 0x26},        // VK_UP
-        {Fw::KeyRight, 0x27},     // VK_RIGHT
-        {Fw::KeyDown, 0x28},      // VK_DOWN
-        {Fw::KeyInsert, 0x2D},    // VK_INSERT
-        {Fw::KeyDelete, 0x2E},    // VK_DELETE
-        {Fw::KeyPause, 0x13},     // VK_PAUSE
-        {Fw::KeyPrintScreen, 0x2C}, // VK_SNAPSHOT
-        {Fw::KeyNumLock, 0x90},   // VK_NUMLOCK
-        {Fw::KeyScrollLock, 0x91},// VK_SCROLL
-        {Fw::KeyCapsLock, 0x14},  // VK_CAPITAL
-        {Fw::KeyCtrl, 0x11},      // VK_CONTROL
-        {Fw::KeyShift, 0x10},     // VK_SHIFT
-        {Fw::KeyAlt, 0x12},       // VK_MENU
-        {Fw::KeyMeta, 0x5B},      // VK_LWIN
-        {Fw::KeyMenu, 0x5D},      // VK_APPS
-        {Fw::KeySpace, 0x20},     // VK_SPACE
-        {Fw::KeyExclamation, 0x31}, // '1' key
-        {Fw::KeyQuote, 0xDE},
-        {Fw::KeyNumberSign, 0x33},
-        {Fw::KeyDollar, 0x34},
-        {Fw::KeyPercent, 0x35},
-        {Fw::KeyAmpersand, 0x37},
-        {Fw::KeyApostrophe, 0xDE},
-        {Fw::KeyLeftParen, 0x39},
-        {Fw::KeyRightParen, 0x30},
-        {Fw::KeyAsterisk, 0x38},
-        {Fw::KeyPlus, 0xBB},
-        {Fw::KeyComma, 0xBC},
-        {Fw::KeyMinus, 0xBD},
-        {Fw::KeyPeriod, 0xBE},
-        {Fw::KeySlash, 0xBF},
-        {Fw::Key0, 0x30},
-        {Fw::Key1, 0x31},
-        {Fw::Key2, 0x32},
-        {Fw::Key3, 0x33},
-        {Fw::Key4, 0x34},
-        {Fw::Key5, 0x35},
-        {Fw::Key6, 0x36},
-        {Fw::Key7, 0x37},
-        {Fw::Key8, 0x38},
-        {Fw::Key9, 0x39},
-        {Fw::KeyColon, 0xBA},
-        {Fw::KeySemicolon, 0xBA},
-        {Fw::KeyLess, 0xBC},
-        {Fw::KeyEqual, 0xBB},
-        {Fw::KeyGreater, 0xBE},
-        {Fw::KeyQuestion, 0xBF},
-        {Fw::KeyAtSign, 0x32},
-        {Fw::KeyA, 0x41}, {Fw::KeyB, 0x42}, {Fw::KeyC, 0x43}, {Fw::KeyD, 0x44},
-        {Fw::KeyE, 0x45}, {Fw::KeyF, 0x46}, {Fw::KeyG, 0x47}, {Fw::KeyH, 0x48},
-        {Fw::KeyI, 0x49}, {Fw::KeyJ, 0x4A}, {Fw::KeyK, 0x4B}, {Fw::KeyL, 0x4C},
-        {Fw::KeyM, 0x4D}, {Fw::KeyN, 0x4E}, {Fw::KeyO, 0x4F}, {Fw::KeyP, 0x50},
-        {Fw::KeyQ, 0x51}, {Fw::KeyR, 0x52}, {Fw::KeyS, 0x53}, {Fw::KeyT, 0x54},
-        {Fw::KeyU, 0x55}, {Fw::KeyV, 0x56}, {Fw::KeyW, 0x57}, {Fw::KeyX, 0x58},
-        {Fw::KeyY, 0x59}, {Fw::KeyZ, 0x5A},
-        {Fw::KeyLeftBracket, 0xDB},
-        {Fw::KeyBackslash, 0xDC},
-        {Fw::KeyRightBracket, 0xDD},
-        {Fw::KeyCaret, 0x36},
-        {Fw::KeyUnderscore, 0xBD},
-        {Fw::KeyGrave, 0xC0},
-        {Fw::KeyLeftCurly, 0xDB},
-        {Fw::KeyBar, 0xDC},
-        {Fw::KeyRightCurly, 0xDD},
-        {Fw::KeyTilde, 0xC0},
-        {Fw::KeyF1, 0x70},  {Fw::KeyF2, 0x71},  {Fw::KeyF3, 0x72},
-        {Fw::KeyF4, 0x73},  {Fw::KeyF5, 0x74},  {Fw::KeyF6, 0x75},
-        {Fw::KeyF7, 0x76},  {Fw::KeyF8, 0x77},  {Fw::KeyF9, 0x78},
-        {Fw::KeyF10, 0x79}, {Fw::KeyF11, 0x7A}, {Fw::KeyF12, 0x7B},
-        {Fw::KeyNumpad0, 0x60}, {Fw::KeyNumpad1, 0x61}, {Fw::KeyNumpad2, 0x62},
-        {Fw::KeyNumpad3, 0x63}, {Fw::KeyNumpad4, 0x64}, {Fw::KeyNumpad5, 0x65},
-        {Fw::KeyNumpad6, 0x66}, {Fw::KeyNumpad7, 0x67}, {Fw::KeyNumpad8, 0x68},
-        {Fw::KeyNumpad9, 0x69}
-    };
-
-    auto it = vkMap.find(key);
-    if(it != vkMap.end())
-        return it->second;
-    return key;
-}
-
-static std::u16string cp1252ToUtf16(const std::string& text)
-{
-    static const char16_t table[32] = {
-        0x20AC, 0xFFFD, 0x201A, 0x0192, 0x201E, 0x2026, 0x2020, 0x2021,
-        0x02C6, 0x2030, 0x0160, 0x2039, 0x0152, 0xFFFD, 0x017D, 0xFFFD,
-        0xFFFD, 0x2018, 0x2019, 0x201C, 0x201D, 0x2022, 0x2013, 0x2014,
-        0x02DC, 0x2122, 0x0161, 0x203A, 0x0153, 0xFFFD, 0x017E, 0x0178
-    };
-
-    std::u16string out;
-    out.reserve(text.size());
-    for(unsigned char ch : text) {
-        if(ch < 0x80)
-            out.push_back(static_cast<char16_t>(ch));
-        else if(ch < 0xA0)
-            out.push_back(table[ch - 0x80]);
-        else
-            out.push_back(0x00A0 + (ch - 0xA0));
-    }
-    return out;
-}
 
 
 UICEFWebView::UICEFWebView()
@@ -1261,189 +1124,77 @@ void UICEFWebView::drawSelf(Fw::DrawPane drawPane)
 
 bool UICEFWebView::onMousePress(const Point& mousePos, Fw::MouseButton button)
 {
-    if (m_browser) {
-        CefRefPtr<CefBrowserHost> host = m_browser->GetHost();
-        if (host) {
-            CefMouseEvent event;
-            Point localPos = mousePos - getPosition();
-            event.x = localPos.x;
-            event.y = localPos.y;
-            event.modifiers = getCefModifiers();
-
-            CefBrowserHost::MouseButtonType btnType = MBT_LEFT;
-            if (button == Fw::MouseRightButton)
-                btnType = MBT_RIGHT;
-            else if (button == Fw::MouseMidButton)
-                btnType = MBT_MIDDLE;
-
-            host->SendMouseClickEvent(event, btnType, false, 1);
-        }
-    }
+    Point localPos = mousePos - getPosition();
+    if (m_browser)
+        CefInputHandler::handleMousePress(m_browser, localPos, button);
 
     return UIWidget::onMousePress(mousePos, button);
 }
 
 bool UICEFWebView::onMouseRelease(const Point& mousePos, Fw::MouseButton button)
 {
-    if (m_browser) {
-        CefRefPtr<CefBrowserHost> host = m_browser->GetHost();
-        if (host) {
-            CefMouseEvent event;
-            Point localPos = mousePos - getPosition();
-            event.x = localPos.x;
-            event.y = localPos.y;
-            event.modifiers = getCefModifiers();
-
-            CefBrowserHost::MouseButtonType btnType = MBT_LEFT;
-            if (button == Fw::MouseRightButton)
-                btnType = MBT_RIGHT;
-            else if (button == Fw::MouseMidButton)
-                btnType = MBT_MIDDLE;
-
-            host->SendMouseClickEvent(event, btnType, true, 1);
-        }
-    }
+    Point localPos = mousePos - getPosition();
+    if (m_browser)
+        CefInputHandler::handleMouseRelease(m_browser, localPos, button);
 
     return UIWidget::onMouseRelease(mousePos, button);
 }
 
 bool UICEFWebView::onMouseMove(const Point& mousePos, const Point& mouseMoved)
 {
-    if (m_browser) {
-        CefRefPtr<CefBrowserHost> host = m_browser->GetHost();
-        if (host) {
-            CefMouseEvent event;
-            Point localPos = mousePos - getPosition();
-            event.x = localPos.x;
-            event.y = localPos.y;
-            m_lastMousePos = localPos;
-            event.modifiers = getCefModifiers();
-            bool leave = !containsPoint(mousePos);
-            host->SendMouseMoveEvent(event, leave);
-        }
-    }
+    Point localPos = mousePos - getPosition();
+    m_lastMousePos = localPos;
+    if (m_browser)
+        CefInputHandler::handleMouseMove(m_browser, localPos, !containsPoint(mousePos));
 
     return UIWidget::onMouseMove(mousePos, mouseMoved);
 }
 
 bool UICEFWebView::onMouseWheel(const Point& mousePos, Fw::MouseWheelDirection direction)
 {
-    if (m_browser) {
-        CefRefPtr<CefBrowserHost> host = m_browser->GetHost();
-        if (host) {
-            CefMouseEvent event;
-            Point localPos = mousePos - getPosition();
-            event.x = localPos.x;
-            event.y = localPos.y;
-            event.modifiers = getCefModifiers();
-
-            int deltaY = direction == Fw::MouseWheelUp ? 120 : -120;
-            host->SendMouseWheelEvent(event, 0, deltaY);
-        }
-    }
+    Point localPos = mousePos - getPosition();
+    if (m_browser)
+        CefInputHandler::handleMouseWheel(m_browser, localPos, direction);
 
     return UIWidget::onMouseWheel(mousePos, direction);
 }
 
 void UICEFWebView::onHoverChange(bool hovered)
 {
-    if (!hovered && m_browser) {
-        CefRefPtr<CefBrowserHost> host = m_browser->GetHost();
-        if (host) {
-            CefMouseEvent event;
-            event.x = m_lastMousePos.x;
-            event.y = m_lastMousePos.y;
-            event.modifiers = getCefModifiers();
-            host->SendMouseMoveEvent(event, true);
-        }
-    }
+    if (!hovered && m_browser)
+        CefInputHandler::handleMouseMove(m_browser, m_lastMousePos, true);
 
     UIWidget::onHoverChange(hovered);
 }
 
 bool UICEFWebView::onKeyDown(uchar keyCode, int keyboardModifiers)
 {
-    if (m_browser) {
-        CefRefPtr<CefBrowserHost> host = m_browser->GetHost();
-        if (host) {
-            CefKeyEvent event = {};  // Initialize all fields to zero
-            event.type = KEYEVENT_RAWKEYDOWN;
-            event.modifiers = getCefModifiers();
-            event.windows_key_code = translateKeyCode(keyCode);
-            event.native_key_code = event.windows_key_code;
-            event.is_system_key = 0;
-            event.character = 0;
-            event.unmodified_character = 0;
-            event.focus_on_editable_field = 0;
-            host->SendKeyEvent(event);
-        }
-    }
+    if (m_browser)
+        CefInputHandler::handleKeyDown(m_browser, keyCode);
 
     return UIWidget::onKeyDown(keyCode, keyboardModifiers);
 }
 
 bool UICEFWebView::onKeyPress(uchar keyCode, int keyboardModifiers, int autoRepeatTicks)
 {
-    if (m_browser && autoRepeatTicks > 0) {
-        CefRefPtr<CefBrowserHost> host = m_browser->GetHost();
-        if (host) {
-            CefKeyEvent event = {};  // Initialize all fields to zero
-            event.type = KEYEVENT_RAWKEYDOWN;
-            event.modifiers = getCefModifiers();
-            event.windows_key_code = translateKeyCode(keyCode);
-            event.native_key_code = event.windows_key_code;
-            event.is_system_key = 0;
-            event.character = 0;
-            event.unmodified_character = 0;
-            event.focus_on_editable_field = 0;
-            host->SendKeyEvent(event);
-        }
-    }
+    if (m_browser && autoRepeatTicks > 0)
+        CefInputHandler::handleKeyPress(m_browser, keyCode);
 
     return UIWidget::onKeyPress(keyCode, keyboardModifiers, autoRepeatTicks);
 }
 
 bool UICEFWebView::onKeyText(const std::string& keyText)
 {
-    if (m_browser && !keyText.empty()) {
-        CefRefPtr<CefBrowserHost> host = m_browser->GetHost();
-        if (host) {
-            std::u16string chars = cp1252ToUtf16(keyText);
-            for (char16_t ch : chars) {
-                CefKeyEvent event = {};  // Initialize all fields to zero
-                event.type = KEYEVENT_CHAR;
-                event.modifiers = getCefModifiers();
-                event.character = ch;
-                event.unmodified_character = ch;
-                event.windows_key_code = static_cast<int>(ch);
-                event.native_key_code = event.windows_key_code;
-                event.is_system_key = 0;
-                event.focus_on_editable_field = 0;
-                host->SendKeyEvent(event);
-            }
-        }
-    }
+    if (m_browser && !keyText.empty())
+        CefInputHandler::handleKeyText(m_browser, keyText);
 
     return UIWidget::onKeyText(keyText);
 }
 
 bool UICEFWebView::onKeyUp(uchar keyCode, int keyboardModifiers)
 {
-    if (m_browser) {
-        CefRefPtr<CefBrowserHost> host = m_browser->GetHost();
-        if (host) {
-            CefKeyEvent event = {};  // Initialize all fields to zero
-            event.type = KEYEVENT_KEYUP;
-            event.modifiers = getCefModifiers();
-            event.windows_key_code = translateKeyCode(keyCode);
-            event.native_key_code = event.windows_key_code;
-            event.is_system_key = 0;
-            event.character = 0;
-            event.unmodified_character = 0;
-            event.focus_on_editable_field = 0;
-            host->SendKeyEvent(event);
-        }
-    }
+    if (m_browser)
+        CefInputHandler::handleKeyUp(m_browser, keyCode);
 
     return UIWidget::onKeyUp(keyCode, keyboardModifiers);
 }
